feat: Add flatpickr for date range selection and integrate with vehicle tracking
- Added flatpickr dependency for date range input in vehicle tracking.
- Implemented date range selection in the tracking map with a default range of the last 7 days.
- Enhanced vehicle detail sidebar with additional information and activity tabs.
- Updated input components to use consistent styling and added a new date input component.
- Improved vehicle search functionality in the sidebar.
- Integrated Select2 for better vehicle selection in the heavy equipment edit form.
- Refactored vehicle status and condition selection for better readability.
- Added CarbonInterval for better time formatting in the project map route.

+++ app/Http/Controllers/HeavyEquipmentController.php
@@ -3,11 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Models\AttendanceLog;
+use App\Models\CartrackVehicle;
 use Illuminate\Http\Request;
 use App\Models\HeavyEquipment;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
+
 class HeavyEquipmentController extends Controller
 {
     public function index(Request $request)
@@ -16,13 +18,13 @@ class HeavyEquipmentController extends Controller
         $kondisi = $request->input('kondisi');
 
         $heavyEquipments = HeavyEquipment::query()
-        ->when($status, function ($query, $status) {
-            return $query->where('status', $status);
-        })
-        ->when($kondisi, function ($query, $kondisi) {
-            return $query->where('kondisi', $kondisi);
-        })
-        ->paginate(10);
+            ->when($status, function ($query, $status) {
+                return $query->where('status', $status);
+            })
+            ->when($kondisi, function ($query, $kondisi) {
+                return $query->where('kondisi', $kondisi);
+            })
+            ->paginate(10);
 
         return view('heavy_equipments.index', compact('heavyEquipments'));
     }
@@ -59,11 +61,11 @@ class HeavyEquipmentController extends Controller
         $heavyEquipment = HeavyEquipment::findOrFail($id);
 
         // Query dimodifikasi untuk menggabungkan data yang sama
-        $hoursMeterHistory = AttendanceLog::whereIn('work_assignment_id', function($query) use ($id) {
-                $query->select('id')
-                    ->from('work_assignments')
-                    ->where('heavy_equipment_id', $id);
-            })
+        $hoursMeterHistory = AttendanceLog::whereIn('work_assignment_id', function ($query) use ($id) {
+            $query->select('id')
+                ->from('work_assignments')
+                ->where('heavy_equipment_id', $id);
+        })
             ->whereNotNull('hours_meter_start')
             ->whereNotNull('hours_meter_end')
             ->select(
@@ -77,7 +79,7 @@ class HeavyEquipmentController extends Controller
             ->groupBy('work_assignment_id', DB::raw('DATE(check_in_time)'))
             ->orderBy('date', 'desc')
             ->get()
-            ->map(function($log) {
+            ->map(function ($log) {
                 return [
                     'date' => Carbon::parse($log->date)->format('d/m/Y'),
                     'start_meter' => $log->start_meter,
@@ -94,13 +96,19 @@ class HeavyEquipmentController extends Controller
     public function edit($id)
     {
         $heavyEquipment = HeavyEquipment::find($id);
+
+        if (!$heavyEquipment) {
+            # code...
+            return redirect()->route('alat-berat.index')->with('error', 'Alat berat tidak ditemukan.');
+        }
+
         $heavyEquipment->load('workAssignments.city', 'workAssignments.district');
-        return view('heavy_equipments.edit', compact('heavyEquipment'));
+        $cartrackVehicles = CartrackVehicle::all();
+        return view('heavy_equipments.edit', compact('heavyEquipment', 'cartrackVehicles'));
     }
 
     public function update(Request $request, $id)
     {
-
         $validatedData = $request->validate([
             'name' => 'required|max:255',
             'nomor_lambung' => 'required|max:255',
@@ -117,6 +125,12 @@ class HeavyEquipmentController extends Controller
         ]);
 
         $heavyEquipment =  HeavyEquipment::find($id);
+
+        if (!$heavyEquipment) {
+            # code...
+            return redirect()->route('alat-berat.index')->with('error', 'Alat berat tidak ditemukan.');
+        }
+
         $heavyEquipment->update([
             'name' => $validatedData['name'],
             'nomor_lambung'  => $validatedData['nomor_lambung'],
@@ -134,6 +148,12 @@ class HeavyEquipmentController extends Controller
     public function destroy($id)
     {
         $heavyEquipment = HeavyEquipment::find($id);
+
+        if (!$heavyEquipment) {
+            # code...
+            return redirect()->route('alat-berat.index')->with('error', 'Alat berat tidak ditemukan.');
+        }
+
         $heavyEquipment->delete();
 
         return redirect()->route('alat-berat.index')->with('success', 'Alat berat berhasil dihapus.');
@@ -143,12 +163,19 @@ class HeavyEquipmentController extends Controller
     {
         return HeavyEquipment::all();
     }
+
     public function updateKondisi(Request $request, $id)
     {
-        $heavyEquipment =  HeavyEquipment::find($id);
         $validatedData = $request->validate([
             'kondisi' => 'required|in:baik,rusak_ringan,rusak_berat',
         ]);
+
+        $heavyEquipment =  HeavyEquipment::find($id);
+
+        if (!$heavyEquipment) {
+            # code...
+            return redirect()->route('alat-berat.index')->with('error', 'Alat berat tidak ditemukan.');
+        }
 
         $heavyEquipment->kondisi = $validatedData['kondisi'];
 
@@ -156,13 +183,13 @@ class HeavyEquipmentController extends Controller
 
         return back()->with('success', 'Kondisi alat berat berhasil diperbarui.');
     }
-  public function getHoursMeterHistory($id)
+    public function getHoursMeterHistory($id)
     {
-        $hoursMeterHistory = AttendanceLog::whereIn('work_assignment_id', function($query) use ($id) {
-                $query->select('id')
-                    ->from('work_assignments')
-                    ->where('heavy_equipment_id', $id);
-            })
+        $hoursMeterHistory = AttendanceLog::whereIn('work_assignment_id', function ($query) use ($id) {
+            $query->select('id')
+                ->from('work_assignments')
+                ->where('heavy_equipment_id', $id);
+        })
             ->whereNotNull('hours_meter_start')
             ->whereNotNull('hours_meter_end')
             ->select(
@@ -174,10 +201,10 @@ class HeavyEquipmentController extends Controller
             ->orderBy('check_in_time', 'desc')  // tetap desc untuk mendapatkan 3 data terbaru
             ->take(3)
             ->get()
-            ->groupBy(function($log) {
+            ->groupBy(function ($log) {
                 return Carbon::parse($log->date)->format('d/m/Y');
             })
-            ->map(function($logs, $date) {
+            ->map(function ($logs, $date) {
                 $maxHoursMeter = $logs->max('hours_meter');
                 return [
                     'week' => $date,
@@ -192,13 +219,12 @@ class HeavyEquipmentController extends Controller
 
     public function getTrackingData($id)
     {
-    
         try {
             // Ambil work assignments terkait
             $workAssignments = DB::table('work_assignments')
                 ->where('heavy_equipment_id', $id)
                 ->pluck('id');
-        
+
             // Query attendance logs
             $attendanceLogs = AttendanceLog::whereIn('work_assignment_id', $workAssignments)
                 ->select(
@@ -210,9 +236,9 @@ class HeavyEquipmentController extends Controller
                 )
                 ->orderBy('check_in_time')
                 ->get();
-    
+
             // Transform data
-            $transformedData = $attendanceLogs->map(function($log) {
+            $transformedData = $attendanceLogs->map(function ($log) {
                 try {
                     if (!$log->check_in_location) {
                         Log::warning("Missing check_in_location for log", [
@@ -221,9 +247,9 @@ class HeavyEquipmentController extends Controller
                         ]);
                         return null;
                     }
-    
+
                     list($lat, $lon) = explode(',', $log->check_in_location);
-                    
+
                     return [
                         'date' => Carbon::parse($log->date)->format('d/m/Y'),
                         'time' => $log->time,
@@ -240,15 +266,15 @@ class HeavyEquipmentController extends Controller
                     return null;
                 }
             })
-            ->filter() // Remove null entries
-            ->values();
-    
+                ->filter() // Remove null entries
+                ->values();
+
             // Group by date
             $groupedData = $transformedData->groupBy('date');
-     
-    
+
+
             // Get last points for each date
-            $trackingData = $groupedData->map(function($group) {
+            $trackingData = $groupedData->map(function ($group) {
                 $lastPoint = $group->last();
                 $result = [
                     'date' => $lastPoint['date'],
@@ -259,21 +285,20 @@ class HeavyEquipmentController extends Controller
                     'visit_count' => $group->count(),
                     'work_assignment_id' => $lastPoint['work_assignment_id']
                 ];
-    
+
                 return $result;
             })
-            ->values();
-    
-    
+                ->values();
+
+
             return response()->json($trackingData);
-    
         } catch (\Exception $e) {
             Log::error("Error retrieving tracking data", [
                 'heavy_equipment_id' => $id,
                 'error' => $e->getMessage(),
                 'trace' => $e->getTraceAsString()
             ]);
-    
+
             return response()->json([
                 'error' => 'Failed to retrieve tracking data',
                 'message' => $e->getMessage()
